uploading changes of the mentioned errors
