+++ package.json
@@ -6,7 +6,7 @@
     "build": "tsc",
     "dev:build": "tsc",
     "dev:serve": "nodemon -e js -w dist ./dist/src/main.js",
-    "start": ": node ./dist/src/main.js"
+    "dev:dev": "nodemon -L --files src/main.ts --exec ts-node"
   },
   "keywords": [],
   "author": "",
@@ -24,12 +24,12 @@
     "node": "16.20.1"
   },
   "devDependencies": {
-    "@types/express": "^4.17.19",
+    "@types/express": "^4.17.21",
     "@types/jsonwebtoken": "^9.0.5",
     "@types/node": "^20.8.4",
     "mongoose": "^6.12.3",
+    "nodemon": "^3.0.2",
     "ts-node": "^10.9.1",
     "typescript": "^5.2.2"
   }
 }
-
