+++ src/__mock__/data.ts
@@ -1,6 +1,4 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
 const data = [
     "test1"
-];
-exports.default = data;
+]
+export default data;
